requirements.txt
torch==2.0.0
torchvision==0.15.1
mlflow==2.3.1
optuna==3.1.1
numpy==1.24.3
Pillow==9.5.0
pandas==2.0.1
scikit-learn==1.2.2
dvc==2.56.0
py-cpuinfo==9.0.0
rich==13.3.5
typer==0.9.0
fastapi==0.95.1
uvicorn==0.22.0
python-multipart==0.0.6
dvc-s3==2.22.0